basicinventory/requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
Django==1.11.3
djangorestframework==3.6.3
idna==2.5
itsdangerous==0.24
MarkupSafe==1.0
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
Werkzeug==0.12.2
